feat: add display scaling
Add ability to handle display scaling. This is specifically for operating system display scales and not for rendering objects in OpenGL. Previously there was an issue with the canvas for OpenGL only covering a portion of the screen if the display was upscaled, now the canvas should fill the whole screen regardless of display scale.