fix: aspect ratio
Add a fixed aspect ratio instead of trying to calculate one based on the screen width. This should fix issues with rendering objects in OpenGL regardless of screen and display size. Also refactor and add clipping plane property to floor plan since this is needed in multiple parts of app.