fix: XY plane movement
Moving in XY plane now works as expected using arrow keys. So if I am turned in a particular direction, the camera moves in that direction (as opposed to just moving with camera turned facing the same direction). Still need to work on moving up/down based on camera tilt...