feat: diffuse materials
Add a simple diffuse material to the spheres. NOTE that this has some really bad "shadow acne", which creates some weird circles in the spheres. This is slightly visible in low res images, but very noticeable in high res ones. Correcting shadow acne does fix this, but I'm wondering if sticking with floats instead of doubles is causing this weird behavior since it doesn't seem like these are in the sample images in the book used to create ray tracer...