fix: randomness
Use thread safe random number generator since ray tracer is run in it's own thread. This seems to fix issue with random numbers not being so random before as the final scene in the ray tracer now has a mix of small spheres (diffuse, metal, and glass), and the spheres are actually randomly placed in the world (instead of displaying in lines).