fix: issues rendering cubes
Add all quads from a cube to world so they are treated separately when determining IntersectRecord. This fixes issue rendering cubes where some sides of cube appear to be missing due to assumption that an intersect occurs for a cube if any of its quads are intersected. That assumption is wrong...